Add files via upload
Pre-requisites
Some experience working with Python
Basic idea about nodes, clients, and servers. And the fact that there may be several server applications and several client applications on a node.
Some idea about MAC addresses and IP addresses, mainly the fact how your computer has a globally unique MAC address but it’s IP address is dependent on the network it is connected to, currently.